+++ lib/src/main/java/com/smileidentity/compose/SmileIDExt.kt
@@ -56,6 +56,8 @@ import kotlinx.collections.immutable.persistentMapOf
  * front camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for SmartSelfie.
+ * @param useStrictMode Whether to use strict mode or not. Strict mode enables an advanced, more
+ * secure, and more accurate UX for higher pass rates. [showInstructions] will be ignored
  * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
@@ -72,16 +74,15 @@ fun SmileID.SmartSelfieEnrollment(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
+    useStrictMode: Boolean = false,
     extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
 ) {
-    // TODO: Move this to a function parameter once we decided to expose it
-    val useExperimentalUi = false
-    val useStrictMode = false
     MaterialTheme(colorScheme = colorScheme, typography = typography) {
-        if (useExperimentalUi) {
+        // TODO: Eventually use the new UI even for nonStrictMode, but with active liveness disabled
+        if (useStrictMode) {
             val context = LocalContext.current
             val selfieQualityModel = remember { SelfieQualityModel.newInstance(context) }
             OrchestratedSelfieCaptureScreenV2(
@@ -129,6 +130,8 @@ fun SmileID.SmartSelfieEnrollment(
  * front camera will be used.
  * @param showAttribution Whether to show the Smile ID attribution or not on the Instructions screen
  * @param showInstructions Whether to deactivate capture screen's instructions for SmartSelfie.
+ * @param useStrictMode Whether to use strict mode or not. Strict mode enables an advanced, more
+ * secure, and more accurate UX for higher pass rates. [showInstructions] will be ignored
  * @param extraPartnerParams Custom values specific to partners
  * @param colorScheme The color scheme to use for the UI. This is passed in so that we show a Smile
  * ID branded UI by default, but allow the user to override it if they want.
@@ -145,16 +148,15 @@ fun SmileID.SmartSelfieAuthentication(
     allowAgentMode: Boolean = false,
     showAttribution: Boolean = true,
     showInstructions: Boolean = true,
+    useStrictMode: Boolean = false,
     extraPartnerParams: ImmutableMap<String, String> = persistentMapOf(),
     colorScheme: ColorScheme = SmileID.colorScheme,
     typography: Typography = SmileID.typography,
     onResult: SmileIDCallback<SmartSelfieResult> = {},
 ) {
-    // TODO: Move this to a function parameter once we decided to expose it
-    val useExperimentalUi = false
-    val useStrictMode = false
     MaterialTheme(colorScheme = colorScheme, typography = typography) {
-        if (useExperimentalUi) {
+        // TODO: Eventually use the new UI even for nonStrictMode, but with active liveness disabled
+        if (useStrictMode) {
             val context = LocalContext.current
             val selfieQualityModel = remember { SelfieQualityModel.newInstance(context) }
             OrchestratedSelfieCaptureScreenV2(